[upgrade] ITankのBulletIntervalの型を変更

+++ Assets/Script/ITank.cs
@@ -4,7 +4,7 @@ public interface ITank
     public int Hp { get; }
     public int ATK { get; }
     public int MoveSpeed { get; }
-    public int BulletInterval { get; }
+    public float BulletInterval { get; }
     [PunRPC]
     public void Hit(int atk);
     public void Die();
